requirements.txt
python-telegram-bot==20.5
langchain==0.0.292
faiss-cpu==1.7.4
python-dotenv==1.0.0
sentence-transformers==2.2.2
llama-index==0.8.28